--- src/main/resources/META-INF/dubbo/com.alibaba.dubbo.registry.RegistryFactory
@@ -1,1 +1,0 @@
-nacos=com.alibaba.dubbo.registry.nacos.NacosRegistryFactory
