+++ src/main/resources/META-INF/dubbo/org.apache.dubbo.registry.RegistryFactory
@@ -1,0 +1,1 @@
+nacos=org.apache.dubbo.registry.nacos.NacosRegistryFactory
