+++ src/test/java/org/apache/dubbo/registry/nacos/NacosRegistryTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.registry.nacos;
+package org.apache.dubbo.registry.nacos;
 
 /**
  * {@link NacosRegistry} Test
