+++ src/test/resources/META-INF/spring/dubbo-consumer-context.xml
@@ -11,6 +11,6 @@
     <dubbo:registry address="nacos://127.0.0.1:8848"/>
 
     <!-- 引用服务接口 -->
-    <dubbo:reference id="demoService" interface="com.alibaba.dubbo.demo.service.DemoService" version="2.0.0"/>
+    <dubbo:reference id="demoService" interface="org.apache.dubbo.demo.service.DemoService" version="2.0.0"/>
 
 </beans>
